Allow multiple requests on one route
Some tests need multiple requests and using plugins is not that
handy if you want to use preconfigured routes.

If you do not provide an explicit option for that, it will
set the maxRequests to Infinity. Before there was a `many`
method in hock it had a hard limit of 1.

Additionally, options are now passed 1:1 to hock, if they are an
object