remove broken wildcard route for /missing
remove catch-all-route, as we return the hock-server instance that
users can work with to add special routes if needed, like
`missing`.

as we return hock itself and the user can add all kinds of routes
that e.g. just match with a special body the feature got impossible
and is also superfluous looking at the current usage pattern
seen in npm tests